src/trees/binaryTreeBreathFirstSearch.test.ts
import search from "./binaryTreeBreathFirstSearch";
import { BinaryTreeNode } from "../types";

const tree: BinaryTreeNode<number> = {
  value: 20,
  right: {
    value: 50,
    right: {
      value: 100,
      right: null,
      left: null,
    },
    left: {
      value: 30,
      right: {
        value: 45,
        right: null,
        left: null,
      },
      left: {
        value: 29,
        right: null,
        left: null,
      },
    },
  },
  left: {
    value: 10,
    right: {
      value: 15,
      right: null,
      left: null,
    },
    left: {
      value: 5,
      right: {
        value: 7,
        right: null,
        left: null,
      },
      left: null,
    },
  },
};

test("Binary tree breath first search", () => {
  expect(search(tree, 45)).toEqual(true);
  expect(search(tree, 7)).toEqual(true);
  expect(search(tree, 69)).toEqual(false);
});
